+++ src/TestSuiteStorage/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add admin UI action fixture class `\Heptacom\HeptaConnect\TestSuite\Storage\Fixture\AdminUiAction\StorageTestUiAction` to provide a basic admin UI action
+- Add `\Heptacom\HeptaConnect\TestSuite\Storage\Action\UiAuditTrailTestContract` to test basic life cycle of logging a UI action
+
 ### Changed
 
 ### Deprecated